moved load method to NerBaseModel
